ibc_apps::nft_transfer::handler
Function process_recv_packet_execute

pub fn process_recv_packet_execute<Ctx>(
    ctx_b: &mut Ctx,
    packet: &Packet,
    data: PacketData,
) -> Result<ModuleExtras, Box<(ModuleExtras, NftTransferError)>>
where
    Ctx: NftTransferExecutionContext,

This function handles the transfer receiving logic.

Note that send/mint_nft_validate steps are performed on the host chain to validate accounts and NFT info. But the result is then used for execution on the IBC side, including storing acknowledgements and emitting events.