ibc_apps::transfer::context
Trait TokenTransferExecutionContext

pub trait TokenTransferExecutionContext: TokenTransferValidationContext {
    // Required methods
    fn escrow_coins_execute(
        &mut self,
        from_account: &Self::AccountId,
        port_id: &PortId,
        channel_id: &ChannelId,
        coin: &Coin<PrefixedDenom>,
        memo: &Memo,
    ) -> Result<(), HostError>;
    fn unescrow_coins_execute(
        &mut self,
        to_account: &Self::AccountId,
        port_id: &PortId,
        channel_id: &ChannelId,
        coin: &Coin<PrefixedDenom>,
    ) -> Result<(), HostError>;
    fn mint_coins_execute(
        &mut self,
        account: &Self::AccountId,
        coin: &Coin<PrefixedDenom>,
    ) -> Result<(), HostError>;
    fn burn_coins_execute(
        &mut self,
        account: &Self::AccountId,
        coin: &Coin<PrefixedDenom>,
        memo: &Memo,
    ) -> Result<(), HostError>;
}

Methods required in token transfer execution, to be implemented by the host.

Required Methods

fn escrow_coins_execute( &mut self, from_account: &Self::AccountId, port_id: &PortId, channel_id: &ChannelId, coin: &Coin<PrefixedDenom>, memo: &Memo, ) -> Result<(), HostError>

Executes the escrow of the tokens in a user account.

memo field allows incorporating additional contextual details in the escrow execution.

fn unescrow_coins_execute( &mut self, to_account: &Self::AccountId, port_id: &PortId, channel_id: &ChannelId, coin: &Coin<PrefixedDenom>, ) -> Result<(), HostError>

Executes the unescrow of the tokens in a user account.

fn mint_coins_execute( &mut self, account: &Self::AccountId, coin: &Coin<PrefixedDenom>, ) -> Result<(), HostError>

Executes minting of the tokens in a user account.

fn burn_coins_execute( &mut self, account: &Self::AccountId, coin: &Coin<PrefixedDenom>, memo: &Memo, ) -> Result<(), HostError>

Executes burning of the tokens in a user account.

memo field allows incorporating additional contextual details in the burn execution.